fixing refactor missing classes
